requirements.txt
numpy~=1.26.4
networkx~=3.4.2
matplotlib~=3.9.2
ipython~=8.29.0
pillow~=10.4.0
torch~=2.5.1+cu118
torchvision~=0.20.1+cu118
scikit-learn~=1.5.2
black~=24.10.0
bpy~=4.3.0
neptune~=1.13.0